refactor(niri): remove niri feature, better error handling
