+++ src/clients/mod.rs
@@ -14,8 +14,6 @@ pub mod lua;
 pub mod music;
 #[cfg(feature = "network_manager")]
 pub mod networkmanager;
-#[cfg(feature = "niri")]
-pub mod niri;
 #[cfg(feature = "sway")]
 pub mod sway;
 #[cfg(feature = "notifications")]
